More work on #28
I've got get-uptime now using the same invoke-command -asjob (with a 10 sec timeout) method as get-perfcounters and get-rebootstatus.

This necessitated small changes in a couple of other functions.

Get-uptime now runs first, and if it marks the system as DOWN, then get-perfcounters and get-rebootstatus do not run. 

Average runtime is now about 5 secs for each host that is UP, 10 secs for each DOWN host, and less than 2 secs for all other code.