resolves #38
New-ComputerRecord was creating 'events' entries with a zero as the starting point. This would make too much work for Get-Eventlog which would then timeout (t takes longer, the father back in logs it must go). The fix was to set the initial LastSystemEvent and LastApplicationEvent entries to a timestamp one hour prior to now.